+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.29.1] - 2025-08-01
+
 ### Fixed
 
 - libcnb:
@@ -390,7 +393,8 @@ version number. See the changelog below for other changes.
 
 - Remove support for legacy BOM. Remove `Launch::bom`, `Build::bom`, `bom::Bom`, `bom::Entry`. ([#489](https://github.com/heroku/libcnb.rs/pull/489))
 
-[unreleased]: https://github.com/heroku/libcnb.rs/compare/v0.29.0...HEAD
+[unreleased]: https://github.com/heroku/libcnb.rs/compare/v0.29.1...HEAD
+[0.29.1]: https://github.com/heroku/libcnb.rs/compare/v0.29.0...v0.29.1
 [0.29.0]: https://github.com/heroku/libcnb.rs/compare/v0.28.1...v0.29.0
 [0.28.1]: https://github.com/heroku/libcnb.rs/compare/v0.28.0...v0.28.1
 [0.28.0]: https://github.com/heroku/libcnb.rs/compare/v0.27.0...v0.28.0